Comenzando la modularizacion
