+++ funciones.h
@@ -1,0 +1,6 @@
+#ifndef FUNCIONES_H
+#define FUNCIONES_H
+
+int openDirectory(const char *ruta, const char *id, const char *command, char *infoMem);
+
+#endif // FUNCIONES_H
